Lock the state file during deployment actions
Any action that reads and then writes the state file must acquire an
exclusive lock on the state file for the duration of the action.  This
ensures isolation.  Action that only read the state file once don't
need to acquire the lock (since the state file is updated atomically).

Fixes #24.