Automatically set up encrypted links between machines
A Charon machine can set the configuration option
‘deployment.encryptedLinksTo’ to specify a set of machines to which an
encrypted peer-to-peer link should be created automatically.  This
generalizes the hacky support for inter-region tunnels in the EC2
backend.  For instance, if machine "proxy" has

   deployment.encryptedLinksTo = [ "backend1" "backend2" ];

then a VPN link over SSH will be automatically set up between "proxy"
and "backend1", and "proxy" and "backend2".  /etc/hosts is set up so a
machine name refers to the encrypted link if available, or the
unencrypted link otherwise; "<NAME>-unencrypted" and
"<NAME>-encrypted" refer to the unencrypted and encrypted links,
respectively.

Fixes #33.