Ensure idempotency in creating EC2 instances
This ensures that we don't "lose" instances if ‘charon deploy’ is
interrupted after EC2 creates the instance but before we register it
in the database.  Either the user runs ‘charon deploy’ again, in which
case EC2 returns the same instance due to the client token, or the
user runs ‘charon destroy’, in which case we destroy the instance by
looking it up by client token.

LRM-9563