hetzner: Bootstrap Nix(OS) base system.
This is basically what nixos-install does minos the parts we don't need, like
evaluating the nixos configuration on the machine.

Essentially, this only copies tar "closures" of bash and nix onto the machine's
chroot and creates wrappers for it to properly do chroot into /mnt and execute
the corresponding commands from Deployment as if it were the real system.

Substitutes do not work yet, so we're temporarily returning True from
has_really_fast_connection() until it's fixed.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>