hetzner: Use rescue bootstrapper for mounting.
Move most of the stuff we used to have in _install_base_system() into
the new method _bootstrap_rescue(), which only does partitioning if the
install argument is set to true. This makes it possible to use the same
routine for "nixops reboot --rescue" as well as for installation of the
machine.

The downside is, that at the moment the full bootstrap installer is
moved to the rescue system, even in case we already have a fully
installed NixOS laying around. We already have a task for that so it is
on our roadmap and will be fixed soon[TM]. :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>