hetzner: Only do a hard reset if necessary.
Hard reboot is only necessary if we are neither in state UP nor in state
RESCUE, which means, that the machine has gone past lunch and it's okay
to virtually push the reset button.

If we are going to install, we're always using hard reset, as we don't
even know the current state.

For normal rebooting, we don't use self.reboot here, because it's going
to put us in self.STARTING state, which we want to avoid when booting
into rescue system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>