scripts/nixops: Propagate exit code in op_ssh.
Whenever the command failed nixops returned an exit signal != 0, but
that was because an exception was thrown. Now, we set check=False and
correctly propagate the exit signal.

Unfortunately, to get this working, so this fixes a related bug in
ssh_util as well, which tried to pass the check keyword argument to
subprocess.call().

Signed-off-by: aszlig <aszlig@redmoonstudios.org>