deployment: Make extraHosts more deterministic.
The value of extraHosts was quite non-deterministic before and the order
was based on the hashing implementation of Python's dictionaries. Now,
we not only sort it in a more deterministic way but also ensure that the
canonical hostname comes as the first item after the IP address.

In my test deployment, this turns the following extraHosts:

5.6.7.8 fot
127.0.0.1 pik-encrypted
9.0.1.2 pik
1.2.3.4 zoq-unencrypted
9.0.1.2 pik-unencrypted
3.4.5.6 slylandro
1.2.3.4 zoq
5.6.7.8 fot-unencrypted
3.4.5.6 slylandro-unencrypted

Into a more readable variant (sorting is based on canonical host):

5.6.7.8 fot fot-unencrypted
127.0.0.1 fot-encrypted
9.0.1.2 pik pik-unencrypted
3.4.5.6 slylandro slylandro-unencrypted
1.2.3.4 zoq zoq-unencrypted

Signed-off-by: aszlig <aszlig@redmoonstudios.org>