hetzner: Access own subnet through the gateway.
At Hetzner, you always need the gateway in order to access the own
subnet, so we're changing the route for our own subnet to use the
gateway here.

This also calculates subnets on our part now, because in the long term
we want to get rid of acquiring network information from the rescue
system and use the Robot's API to check IP addresses and subnets.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>