tests/hetzner: Directly add some debs to rescue.
So far, there was only openssh, that has been added to the rescue system
itself by adding a specific package-lists hook. It doesn't suffice to
add it just to the dependency list of live-build itself, because it
might not end up being installed in the resulting system. And we now
have even more packages that we need in the resulting system.

That's why we now have a new attribute called additionalRescuePackages,
which is solely for this purpose, so everything that's put in there will
end up being in the live ISO.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>