Add basic unit tests for the new SSH class.
We're going to refactor the class so let's begin by providing some basic
test cases and work on the implementation.

Obviously this makes running the test cases using setup.py fail and thus
the Hydra build product won't succeed as well.

In this new test case we're setting up a small dummp SSH server using
Paramiko, connect to the server and try various commands. Of course at
the moment, we only have two tests, because we need to replace
SSH.run_command().

Signed-off-by: aszlig <aszlig@redmoonstudios.org>