MachineState: Refactor copy_closure_to().
This now should not need the Perl version anymore and will use calls to
nix-store directly instead of passing OpenSSH flags to an external
program.

Of course, this is only a basic version and it won't print the paths as
they're moved onto the target machine, but let's improve that later.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>