ssh_util: Restructure and refactor SSH class.
The SSH class now only is one main class with a connect() method that's
returning a SSHConnection instance. Of course this looks pointless right
now, but that way we can easily implement connection pooling later,
without those dirty hacks like register_*_fun(), which add unnecessary
references to the SSH instance and thus could prevent it from being
garbage collected.

In addition, we now no longer need the SSHMaster process stuff anymore
(which is one of the primary reasons for switching to Paramiko) and the
SSHConnection stays open for further requests until it is either closed
directly or by the pooling functionality (as soon as it's implemented).

Now the downside is that we need to pass more arguments tu methods of
SSHConnection, but that's actually a good thing because we can clearly
separate it from the rest of the code.

And of course, this implementation is neither correctly working at the
moment, nor is it complete. So we have more work ahead.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>