nix_expr: Sort attributes based on their keys.
It's critical not only for the tests that we always get the same output
for the same input. For generating the physical spec this is even more
important, because we want to avoid rebuilding the system configurations
every time a dict want's to be ordered differently than the time before.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>