nix_expr: Support pretty-printing functions.
This adds a new container instance called Function, which takes the
function head as a string value (we don't allow any native Python types
here, just a raw string value) and the body as an arbitrarily nested
Python data type. That should make it possible to pretty-print full
machine configurations.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>