nix_expr: Create a preliminary nix2py().
Although I'd love to use Haskell's Parsec or at least some lightweight
implementation of it (I even started one, consisting of a few hundred
lines of code), I decided to go to for a "poor mans parser combinator"
iplementation. Currently it's just able to parse strings, which is done
with a single regular expression for both string forms (well, the third
form would be an URI, but let's leave that out until we really need it).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>