nix_expr: Implement parsing of attribute sets.
So, this is the most important datatype we need to have, because it's
the most essential one in Nix. Due to the postprocessing of the parsed
list of nested dictionaries, we now also have a function called
merge_dicts() which will come in handy when we implement it for the
physical specification.

What's still missing are functions and "all other kinds of raw values".

Signed-off-by: aszlig <aszlig@redmoonstudios.org>