nix_expr: Implement parsing of RawValue's.
This is rather fragile and might cause parsing to fail for example if
the expression has nested parentheses. But in our case we prefer failing
instead of consuming too much input.

And of course, our goal is to implement only a subset of the Nix
expression language, so we don't need to aim for completeness in any
way.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>