deployment: Integrate pretty printer.
Now, this should switch almost everything from accumulating lines to
using our new pretty-printer. Within the Hetzner backend we also had to
use the parser, because the outputs of nixos-hardware-scan and nixpart
are Nix expressions. So using a simple regex won't help us there,
because the output of these programs might change.

The structure of the get_physical_spec() method of Deployment still is
inherently impure, even though we now have a way more elegant solution
using the pretty printer.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>