deployment: Show individual tracebacks on --debug.
This now ensures, that a full traceback is printed whenever an exception
is raised by a specific resource, because that was no longer the case
after 7cf4cc7f5234a41e057eee3dbcd8e9b8dadd0425.

Also, this prints those exceptions/tracebacks in bright red now, so it's
easier to spot in large deployments.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>