ssh_util: Rewrite event loop in invoke_shell().
We're now setting non-blocking mode for sys.stdin, and resetting it on
cleanup/exit. This especially makes input escape sequences much more
responsive. Also, we're now reading on a per-char basis from the
channel, which works around buffering issues and needs to be properly
fixed later. But for now it does what we want.

Also removing the comment referencing Paramiko's demo code, because
we're already differing from the code and the difference will increase
during the work on the Paramiko integration.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>