ssh_util: Fix buffering issue on early EOF.
The fix is not really very nice, but we're going to refactor it anyway.

But the hardest part here was to properly write a unit test for this,
because we need to send a buffer large enough for the race condition to
happen. Also, we need to shutdown the write end as soon as possible, so
we exactly mimic the behaviour of OpenSSH.

Unfortunately I fear that in some cases (on veeery slow systems) the
test might succeed even without the fix(!), because we are concurrent
here by use of threading for server/client. In any case, I've put my
system under heavy load to make sure this won't happen, so I suppose the
test should be good enough.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>