tests/hetzner: Use Debian Wheezy in rescue system.
As Hetzner updated their rescue system to Debian Wheezy, we also need to
change the virtual CD image of the VM test. In order to correctly work,
we need to ensure that GPG is properly ignored and that additional
required packages for Wheezy are installed as well.

One exception is net-tools, which is no longer required in Wheezy and we
should also use iproute instead of ifconfig, but as long as Debian
doesn't drop ifconfig entirely, we're going to leave it as is and clean
up later.

Thanks again to @lovek323 for reporting this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>