paramiko: Add patch for ECDSA private keys.
This overrides the paramiko derivation to include a patch that allowsa
to use ECDSA private keys, which is the default in the current version
of NixOps.

Previously I've added the patch in NixOS/nixpkgs@58fdf34, but as it is
largely NixOps specific, I revented it und thus added it here directly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>