Add deployment.useHostNixStore option
With deployment.useHostNixStore set to tru, no closures
are copied to the deployed machine. This is useful if
the Nix store is shared from the machine running nixops
to the deployed machine in some way. Currently, useHostNixStore
is only allowed for deployments of type "none", and you have
to make sure the host store is mounted correctly on the
deployed machine.