backends: Check units with systemctl --no-legend.
If legend is enabled, the unit names are indented by two spaces in
systemd 212. But if used with --no-legend we not only get something that
resembles the old behaviour, we also can be sure that we don't match
anything in the legend and/or additional descriptions around the actual
content.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>