tests/none: Properly propagate store dependencies.
Because the coordinator machine needs to build the test machines, we
need to make sure that every package that can't be built without
networking is already available.

So far, the test has used environment.systemPackages to add those
references. However, using pkgs.stdenv won't export a single reference
to the store path of stdenv, because it doesn't contain any paths that
are in the default environment.pathsToLink.

This usually shouldn't be a problem, because stdenv had a
nix-support/propagated-user-env-packages file, which made sure that
every package listed there was recursively added to the system path as
well.

Unfortunately, since NixOS/nixpkgs@328f7a6, stdenv doesn't contain
propagatedUserEnvPkgs anymore, so these dependencies aren't available to
the coordinator machine anymore.

Either way, if any of the packages listed in environment.systemPackages
do not contain paths that are not in environment.pathsToLink, we don't
get any reference to the path and that in turn might lead to unexpected
behaviour.

One example I can think of is if /lib is removed from pathsToLink, which
is already marked as "FIXME: remove" since NixOS/nixpkgs@e636e0a.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>