tests/hetzner: Fix prepopulation of the Nix store.
This is in par with 2a76f9c95795c27f53c1a26cd1167f5170718958.

Over there, we do the same fix for the "none" backend. The approach here
doesn't suffer from the same problems because we were using myEnvFun,
which happens to properly reference the store paths.

While this is working, it's ugly as hell because myEnvFun could not only
cease to exist but also change implementation, so we're now doing the
same as in the referenced commit now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>