tests/none: Don't try to remount /nix/store.
We have virtualisation.writableStore active for the targets that are not
yet deployed by NixOps but not within the NixOps network specification.

This has worked regardless in the past, because using a writable store
was implemented solely using boot.initrd.postMountCommands, so switching
to a new configuration didn't change anything and left the store
untouched.

But after NixOS/nixpkgs@e68b0c7, the mounts for the writable store are
now using the fileSystems configuration attribute, which causes the
mount points to be changed on switch.

Another problem is that the mount units weren't escaped properly. But
that isn't the problem here anymore (because we don't change the mounts
between activation) and has been (or is going to be) dealt with in
NixOS/nixpkgs#10546.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>