tests/none: Properly set IP addresses for eth1.
This has never been done in this test but it's actually done in
<nixpkgs/nixos/lib/build-vms.nix> using assignIPAddresses.

The reason this hasn't been failing was that the IP address
reconfiguration wasn't done properly back then when this test was
happily succeeding. The implementation in <nixpkgs> actually didn't stop
the interface at all, which is why this passed unnoticed.

Since then, we also have proper networking tests in <nixpkgs>, so we
shouldn't assume a broken implementation in <nixpkgs>.

However, the implementation is still a bit ugly, because we're now
statically setting the IP address instead of reusing assignIPAddresses.

We can't reuse the latter, because it requires us to insert a dummy node
(coordinator), which we can't remove later on because the node name is
referenced in the configuration via getAttrs on the node attribute.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>