ssh_util: Reconnect on dead SSH master socket.
This took me hours to debug, because I was originally hunting for the
issue behind the test failure of the Hetzner backend:

https://hydra.nixos.org/build/27122260/nixlog/19/raw

It did not fail on my own Hydra instance or my build machines, because
the build of the target machines was fast enough so the gap between the
last connection to the master and the next connection when copying the
closure to the target machines was less than 600 seconds.

Those 600 seconds is the ControlPersist timeout for the SSH master
connection and was introduced in bbe469a. On a slow or highly loaded
system however, this gap is too long, the master socket dies in-between
and a SSH client connecting to the non-existing sockets waits forever.

Fortunately, OpenSSH cleans up the socket, so we can easily test whether
the socket is still available.

Also, OpenSSH only creates the socket once the connection is
established, so while we usually start the SSHMaster when we want to
connect anyway, it's not necessarily the case.

So we wait 60 seconds until the master socket is created, so that
is_alive() doesn't return False in-between. Which in turn would lead to
the startup of another master process while the first one is still
starting up.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>