ssh_util: Add invoke_shell() to bypass SSHMaster.
Or rather a spawn() method within the SSHMaster class to do direct
connections to the host.

The reason for this is that the current implementation of the ssh
subcommand doesn't use all options that can be available for a
particular machine.

One example would be password authentication while having a machine in
the Hetzner rescue system. Currently we do not correctly pass the
askpass helper to SSH and the connection just fails while the system is
in rescue mode.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>