scripts/nixops: Switch to using SSH.invoke_shell.
We now should have a ssh subcommand which works the same way as the
ssh-for-each subcommand, except that it doesn't rely on a SSH master
socket to be present.

This not only has the benefits of making the codebase more DRY but also
things like ssh into a machine that's in rescue mode now work properly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>