tests/hetzner: Test if SSH works in rescue system.
This actually tests the implementation done in 6781c0a and e856016.

We want to make sure that even when we have to rely on a askpass helper,
the ssh subcommand still works.

So getting this to work with our current tests it also means that we
need to generalize the setupAndStartRescue function to wrap a function
that's solely for the post-boot commands executed on the rescue system,
which we need to reboot into the rescue system *after* the machine has
had its first deployment.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>