tests/hetzner: Improve instrumentation for rescue.
The <nixpkgs/nixos/modules/testing/test-instrumentation.nix> is used as
a reference here, and we want to get as closely as possible to that so
we have the full testing API at hand even for the rescue system.

This also gets rid of the annoying getty at the serial console.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>