deployment: Fix state_version warning for RESCUE.
Hetzner machines are deployed using *only* an (almost) empty store
directly from the rescue system. With *only* I mean there is no /etc (or
to be more exact: only a dummy /etc), so getting the value of
/etc/os-release will not only fail but also won't work at all because
there is no pre-existing deployment on the system.

As 877158c introduced a warning if NixOps is unable to determine the
state_version, we simply shut up the warning whenever we're in rescue
state.

An alternative would be to not even try to get the version when in
RESCUE, but in cases where the deployment is done to a pre-existing
rescue system with a custom kickstart script it's actually legit to
fetch the version even in RESCUE state.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>