scripts/nixops: Use m.ssh.run_command in op_ssh().
We already had that before e57686c2d52813ac8afaaed57a4d951339825ca4.

The reason this has been changed was to be able to supply a user name,
which I had accidentally dropped much earlier in 9cca2d8.

So the actual problem here is that we can't ssh into a machine using
passwords and _not_ have a SSH master. This is because the SSH_ASKPASS
helper is only called when there is no TTY available, which is kind of a
contradiction to what the "nixops ssh" command should actually do.

Of course, we could even manage to do it without a master socket, but
that would involve creating a PTY and pipe everything to SSH forth and
back (and we also would need a bit of signal handling). Which pretty
much boils down to increasing the surface for bugs.

Instead of just reverting back to the old version, we actually now pass
the username along using OpenSSH's -l flag.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>