ssh_util: Add new split_openssh_args() method.
This is useful if we want to distinguish from a bunch of SSH flags
whether the argument is a normal OpenSSH flag or whether it's an actual
command that's passed to OpenSSH.

We do this by simply recursing through the list of flags and eat up all
flags and non-flags until we either get a command or a "--" in which
case we treat everything afterwards as a command.

Note that this is specifically for use in NixOps, because OpenSSH not
only has a command argument but also a target argument which we don't
need so we don't handle it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>