scripts/nixos: Fix handling SSH args in do_ssh.
Regression introduced by da1a81a82936cdc63749b2950674daadf7a1d278.

This really doesn't pass the username to the SSH master process but just
the ssh call that's using the master, which then will obviously fail.

The last two commits were a preparation for improving this, so on one
side we have the user keyword argument, which passes the user name
directly as user@machine to the SSH master process.

On the other side we now have split_openssh_args(), which we can use to
distinguish between a SSH command and a SSH flag. The reason this is
necessary is that we don't want to add the command to the SSH master
process but of course we want to have flags appended to the master
process.

While this certainly increases complexity, but if we really want to pass
SSH options verbatim, we don't have any other option except adding them
like this:

nixops ssh --ssh-flags="-i /foo/bar" my_shiny_command

However, this would introduce a backwards-incompatibility, because SSH
flags passed verbatim will then throw an error.

And I've stated the reason why we can't avoid starting an SSH master
process in da1a81a82936cdc63749b2950674daadf7a1d278 already.

On the upside we gain one more thing (except fixing the functionality)
and it should be mostly backwards-compatible:

We can finally do the following: nixops ssh shiny_machine ls -l

The reason for this is that we now use argparse.REMAINDER, so we don't
get nasty quirks and escaping goofs if "--" is used.

And similar to argparse.REMAINDER is the actual OpenSSH argument parsing
where after the last non-flag the target and the command are used
without trying to parse additional arguments.

The reason why it's only "mostly" backwards-compatible is that something
like this:

nixops ssh foo -- -v ls -l

Has now to be called like this:

nixops ssh foo -v ls -l

Signed-off-by: aszlig <aszlig@redmoonstudios.org>