backends/none: don't use _ssh_private_key if its corresponding public key hasn't been deployed yet
This should fix #361.

The boolean attribute none.sshPublicKeyDeployed is added to the
NoneState and defaults to False. It indicates whether the SSH public key
has been deployed to the target. If it's False the get_ssh_flags()
function will not set the corresponding private key as the identity.

none.sshPublicKeyDeployed will be set to True after activating the
configuration. From then on get_ssh_flags() will set the identity to the
corresponding private key allowing nixops to authenticate itself using
the keypair stored in its DB.