Fix #362: chown the key using the uid and gid
Previously the owner of the uploaded key would be set to the specified
user name and group name. However, when the configuration which defines
this user and group is not yet activated the chown operation will fail.

What we now do instead is set the owner of the key using the UID and GID
computed from the configuration.

This does enable a small attack window after the chown UID:GID and
before the activation of the new configuration which adds the new user
and group. If there's an existing user or group with the specified UID
or GID respectively he might be able to read the key during that window.