tests/hetzner/rescue-image: Update backport debs.
Although the Debian stable repositories change very rarely, the backport
repositories however do not. So in the long run we probably need to move
over to create those debs on our own or provide a third-party mirror for
them.

In the short run, this should fix the Hetzner backend tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>