Fix evaluation and build of manual
First of all, optionAttrSetToDocList now has an additional argument for
specifying the internal _module attributes. Unfortunately, toXML
silently ignores function calls so we don't get an error at this point.

In addition to this, we now have to specify type representations if we
create a new option type with mkOptionType, so let's add those to the
NixOps specific option types.

Both of these issues are caused by NixOS/nixpkgs@cad8957.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>