tests/hetzner/repos: Pin GnuPG version to v2.0
I ran into this while trying to run tests prior to merging #564 (merge
commit b09e618ed63142e1f4babb3d7f5aeb9d63e2b744).

The default GnuPG version in <nixpkgs> is 2.1, but the Debian version
used while building the Debian repository is GnuPG version 2.0, so we
need to make sure both versions match at least in major and minor.

The test for the Hetzner backend now still fails, but at least it
doesn't fail anymore because one of its dependencies are failing.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>