tests/hetzner: Fix up coordinator dependencies
As we don't have networking access within tests, we also need to add
pkgs.lib{xml2,xslt}.dev or pkgs.lib{xml2,xslt} as fallback.

I'm using .dev here, because it references all of the other outputs of
the corresponding library as well so we don't need to list all the
possible outputs.

Unfortunately, I'm not quite sure what has led to this requirement,
because I don't want to bisect a whole lot of changes since NixOS
release 16.03 just to find out the origin of this.

Adding these dependencies now causes the test to succeed with NixOS
unstable and release-16.09.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>