Eliminate duplicate SSH flags
`MachineState.run_command()` passes SSH flags to `self.ssh.run_command()`.
However, `self.get_ssh_flags()` is already registered as a `ssh_flag_fun` in the
class `__init__()` function, so `ssh_util.SSH` already uses it to get the flags
when initiating a connection. This lead to the SSH flags being duplicated, which
causes an error for some flags (e.g. the `-J` flag, which can only be specified once).