Support persistent spot requests
This means when the instance gets interrupted, the request is still
valid, and if the termination behaviour is set to "stop", the instance
will start automatically when the request get fulfilled again by AWS.

There's no way to achieve this with Boto2, so I changed the call that
create the spot request to Boto3. I did some refactoring, so the same
call is used when creating both On demand and spot instances.