tests/hetzner: Fix evaluation with qemu-flags.nix
Earlier this year[1], qemu-flags.nix was turned into a function which
takes an attribute with a pkgs argument. So in order to cope with this
and also be compatible with older nixpkgs versions, I added a check on
isAttrs and only if it's not the pkgs argument is passed to it.

I've only tested whether evaluation works but didn't run the test, so it
might still be broken.

[1]: https://github.com/NixOS/nixpkgs/commit/1ce13804978ef361511014d8fc8

Signed-off-by: aszlig <aszlig@nix.build>