Do not suppress rollback exceptions
I'm not sure what the original motivation for this try/except block
was, though it seems to have been added several years ago. I think if
an exception occurs during rollback, it needs to be either handled
(though I'm not sure there's anything we could do with it) or surfaced
to the user; silent failures can only cause problems.

(I don't mind how the failure is presented; if you'd prefer to avoid a
top-level exception here I can re-add the try/except block with an
error message.)