hetzner: Set prefix length to /32 for main address
This was already suggested by @flokli in #1032 and it basically makes
sure that all traffic is sent to the gateway instead of being sent
directly.

The Hetzner Wiki at [1] describes configuration of Debian to include a
"peertopeer" directive in /etc/network/interfaces, which roughly
translates to the command "ip address add ... peer A.B.C.D".

However, after testing with and without the peer keyword of "ip address"
I didn't notice any difference in behaviour in comparison to setting a
plain /32 prefix length as we do now.

Apart from making configuration less complicated, this also gets rid of
a bunch of code we now no longer need, eg. calculating subnet masks or
getting the real prefix length.

Tested on a newly deployed PX61-NVMe.

[1]: https://wiki.hetzner.de/index.php/Netzkonfiguration_Debian/en#Dedicated_Servers

Signed-off-by: aszlig <aszlig@nix.build>