parallel: run_tasks -> List[Result], not List[Optional[Result]]
If any of the tasks raised an exception, we don't return any results.
In the first typing of this function, I saw we were adding None to
the results even if there was an exception. This was just to see
if we were done collecting or not. Changing this to a counter fixes it
and removes the Optional in the return.