ci: init
This adds a `ci.nix` which is able to generate a `ci.yml` for use with
github actions.
Usage: `nix-build ci.nix --no-out-link | bash`

The main idea behind generating the yaml with nix, as opposed to just
writing yaml, is that Nix is easier to write and maintain. It would've
taken a lot more copy-paste to achieve the same done here in raw gh
actions yaml.