Add flake8 linter
This is mainly an addition to avoid import loops by detecting unused
imports early and force them to be removed.

Flake8 also detects other things which may be of interest of us such
as bad variable naming, function complexity (this one is of great
interest).

We opt out of a number of checks that conflict with black's auto
formatting rules.