Fix legacy build with potential downsides
There was some incompatibilities that prevented the build from
completing, I've patched them to make it work, commenting out a few
lines of code (marked in README) - this is likely break GCE and Azure
potentially, but I haven't tested. Works on existing (pre 2.0) nixops
project out of the box (with DigitalOcean as the cloud of choice).